chore: Separate exception handling into its own file. Update tests.
